kellanova_ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
geopandas==0.14.4
shapely==2.0.4
geopy==2.4.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
httpx==0.27.0
sqlalchemy==2.0.30
duckdb==0.10.3
plotly==5.22.0
folium==0.17.0
streamlit-folium==0.20.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1

